make use of the testruns legend and repres. letter
